+++ packages/blade/src/components/Box/getBoxStyles.ts
@@ -56,7 +56,9 @@ const getBoxStyles = ({
   right,
   bottom,
   background,
+  backgroundColor,
   zIndex,
+  borderRadius,
 }: BoxProps & { theme: Theme }): CSSObject => ({
   display,
   flex,
@@ -90,7 +92,9 @@ const getBoxStyles = ({
   maxWidth: maxWidth ? makeSize(maxWidth) : undefined,
   transform,
   background,
+  backgroundColor,
   zIndex,
+  borderRadius: borderRadius ? makeSize(getIn(theme, `border.radius.${borderRadius}`)) : undefined,
 });
 
 export default getBoxStyles;