+++ packages/blade/src/components/Counter/__tests__/Counter.ssr.test.tsx
@@ -1,0 +1,9 @@
+import { Counter } from '../Counter';
+import renderWithSSR from '~src/_helpers/testing/renderWithSSR.web';
+
+describe('<Counter />', () => {
+  it('should limit value with max prop', () => {
+    const { getByText } = renderWithSSR(<Counter value={20} max={10} />);
+    expect(getByText('10+')).toBeTruthy();
+  });
+});