+++ packages/blade/src/components/Icons/iconMap.ts
@@ -26,20 +26,28 @@ import RupeeIconComponent from './RupeeIcon';
 import SearchIconComponent from './SearchIcon';
 import SettingsIconComponent from './SettingsIcon';
 import SlashIconComponent from './SlashIcon';
+import BankIconComponent from './BankIcon';
 import TrashIconComponent from './TrashIcon';
 import AlertTriangleIconComponent from './AlertTriangleIcon';
 import AlertOctagonIconComponent from './AlertOctagonIcon';
+import BookIconComponent from './BookIcon/BookIcon';
 import CheckIconComponent from './CheckIcon';
 import CheckCircleIconComponent from './CheckCircleIcon';
 import RefreshLeftIconComponent from './RefreshLeftIcon';
 import RotateCounterClockWiseIconComponent from './RotateCounterClockWiseIcon';
 import TrendingUpIconComponent from './TrendingUpIcon';
 import TrendingDownIconComponent from './TrendingDownIcon';
+import UserIconComponent from './UserIcon';
 import UsersIconComponent from './UsersIcon';
 import HelpCircleIconComponent from './HelpCircleIcon';
 import ExternalLinkIconComponent from './ExternalLinkIcon';
 import MinusIconComponent from './MinusIcon';
 import MailIconComponent from './MailIcon';
+import SettlementsIconComponent from './SettlementsIcon';
+import ShuffleIconComponent from './ShuffleIcon';
+import TagIconComponent from './TagIcon';
+import ClockIconComponent from './ClockIcon';
+import TransactionsIconComponent from './TransactionsIcon';
 import type { IconComponent } from '.';
 
 // add new icons here to make them visible in storybook
@@ -51,12 +59,15 @@ const iconMap: Record<string, IconComponent> = {
   ArrowUpRightIcon: ArrowUpRightIconComponent,
   AlertTriangleIcon: AlertTriangleIconComponent,
   AlertOctagonIcon: AlertOctagonIconComponent,
+  BankIcon: BankIconComponent,
+  BookIcon: BookIconComponent,
   CheckIcon: CheckIconComponent,
   CheckCircleIcon: CheckCircleIconComponent,
   ChevronDownIcon: ChevronDownIconComponent,
   ChevronLeftIcon: ChevronLeftIconComponent,
   ChevronRightIcon: ChevronRightIconComponent,
   ChevronUpIcon: ChevronUpIconComponent,
+  ClockIcon: ClockIconComponent,
   CloseIcon: CloseIconComponent,
   CreditCardIcon: CreditCardIconComponent,
   DollarIcon: DollarIconComponent,
@@ -79,13 +90,17 @@ const iconMap: Record<string, IconComponent> = {
   RupeeIcon: RupeeIconComponent,
   SearchIcon: SearchIconComponent,
   SettingsIcon: SettingsIconComponent,
+  SettlementsIcon: SettlementsIconComponent,
+  ShuffleIcon: ShuffleIconComponent,
   SlashIcon: SlashIconComponent,
+  TagIcon: TagIconComponent,
+  TransactionsIcon: TransactionsIconComponent,
   TrashIcon: TrashIconComponent,
   RefreshLeftIcon: RefreshLeftIconComponent,
   RotateCounterClockWiseIcon: RotateCounterClockWiseIconComponent,
   TrendingUpIcon: TrendingUpIconComponent,
   TrendingDownIcon: TrendingDownIconComponent,
-
+  UserIcon: UserIconComponent,
   UsersIcon: UsersIconComponent,
 };