+++ packages/blade/CHANGELOG.md
@@ -1,5 +1,82 @@
 # @razorpay/blade
 
+## 5.5.1
+
+### Patch Changes
+
+- 735e370: fix(blade): update peerDependencies to support react v18
+
+## 5.5.0
+
+### Minor Changes
+
+- a094736: feat: expose `onFocus` on `TextInput` and `TextArea`
+- 2c2841a: added transaction icon
+- 46425d3: feat(blade): add ClockIcon
+- 1dd920e: feat(Icons): add BankIcon
+- 227be3d: added tag, shuffle, user, book, and settlements icons
+- e64d7cc: chore: design changes for Badge, Counter, Spinner
+
+### Patch Changes
+
+- ba16503: fix(blade): TextInput clear button state on initial render
+
+## 5.4.3
+
+### Patch Changes
+
+- e6c3ea9: fix(blade): restrict childrens in Card component
+
+## 5.4.2
+
+### Patch Changes
+
+- da470b0: fix: remove `maxWidth` from Badge
+
+## 5.4.1
+
+### Patch Changes
+
+- 7eb6e4c: feat(Code): Use alpha 50 token in Code component's background
+
+## 5.4.0
+
+### Minor Changes
+
+- 5c8005f: feat: add `ProgressBar` component
+
+## 5.3.0
+
+### Minor Changes
+
+- 26baa81: feat(blade): added Card component
+
+## 5.2.1
+
+### Patch Changes
+
+- 9966931: chore: fix dom nesting in form label component
+- e660831: fix: change acceptable BaseInput `type` from `numeric` to `number`
+
+## 5.2.0
+
+### Minor Changes
+
+- d03de10: feat(Alert): update `isFullWidth` to make inline borderless alerts on desktop
+
+  > **Warning**
+  >
+  > `isBorderless` prop is removed and its usage is now replaced by `isFullWidth`. The layout is updated to match the designs and is now centered on desktop resolutions.
+
+  ### Steps for migration:
+
+  ```diff
+  <Alert
+  - isBorderless
+  + isFullWidth
+  />
+  ```
+
 ## 5.1.5
 
 ### Patch Changes